mockApi : restructured mockApiEndPoint and added delays and other params
