+++ packages/nimbus/electron/server/server.ts
@@ -4,4 +4,8 @@ require('dotenv').config();
 
 
 const PORT = process.env.PORT || 5000;
-app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
+function nodeServer() {
+    app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
+}
+
+export {nodeServer}
