+++ packages/nimbus/electron/server/types/json/index.ts
@@ -1,4 +1,4 @@
-import { Fields } from "../mock/mockData";
+import { Fields } from "../mock/mockDataTemplate";
 
 export type JsonObjectType = {
     [key: string]: keyof Fields | (keyof Fields)[] | JsonObjectType;
